# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Courtside TW API
  version: 1.0.0
  description: Contract-first REST API for public reading, reader library, editorial publication and optional provenance/SIWE boundaries.
servers:
  - url: https://api.courtside.tw
    description: Production API origin placeholder; deployment is a later task.
...
security:
  - bffSession: []
  - oidcBearer: []
paths:
...
  /api/v1/editor/issues:
    post:
      tags:
        - Editorial
      operationId: createEditorIssue
      summary: Create an issue draft
      description: Creates a draft issue owned by the editorial workflow.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueDraft"
              examples:
                valid:
                  value:
                    issueId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    issueNumber: 1
                    version: 1
                    title: Issue 1
                    slug: issue-1
                    description: Issue 1
                    coverAssetId: 0190f7b0-7c4b-7e3a-8f12-123456789abd
                    state: DRAFT
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
        "429":
          $ref: "#/components/responses/Problem429"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueDraftInput"
            examples:
              valid:
                value:
                  title: Issue 1
                  slug: issue-1
                  coverAssetId: 0190f7b0-7c4b-7e3a-8f12-123456789abd
      x-idempotent: true
      x-required-roles:
        - EDITOR
    get:
      tags:
        - Editorial
      operationId: listEditorIssues
      summary: List issue drafts
      description: Lists drafts and workflow states visible to the editorial actor.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueDraftPage"
              examples:
                valid:
                  value:
                    items: []
                    page:
                      nextCursor: null
                      limit: 20
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "429":
          $ref: "#/components/responses/Problem429"
      x-required-roles:
        - EDITOR
    patch:
      tags:
        - Editorial
      operationId: patchEditorIssue
      summary: Patch an issue draft
      description: Updates an issue draft with an If-Match optimistic lock.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueDraft"
              examples:
                valid:
                  value:
                    issueId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    version: 2
                    title: Issue 1 revised
                    slug: issue-1
                    state: DRAFT
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
        "429":
          $ref: "#/components/responses/Problem429"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssuePatch"
            examples:
              valid:
                value:
                  issueId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                  changes:
                    title: Issue 1 revised
      x-idempotent: true
      x-optimistic-lock: If-Match
      x-required-roles:
        - EDITOR
  /api/v1/editor/issues/{issueId}/sections:
    parameters:
      - $ref: "#/components/parameters/IssueId"
    get:
      tags:
        - Editorial
      operationId: listEditorIssueSections
      summary: List issue sections
      description: Returns the server-backed, explicitly positioned table of contents for an issue draft.
      security:
        - bffSession: []
        - oidcBearer: []
      responses:
        "200":
          description: Current section ordering.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueSectionCollection"
              examples:
                valid:
                  value:
                    issueId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    issueVersion: 3
                    sections:
                      - sectionId: 0190f7b0-7c4b-7e3a-8f12-123456789abd
                        title: 場邊現場
                        position: 1
                        articleCount: 2
                        version: 0
                      - sectionId: 0190f7b0-7c4b-7e3a-8f12-123456789abe
                        title: 人物與方法
                        position: 2
                        articleCount: 1
                        version: 0
          headers:
            ETag:
              description: Current issue aggregate version.
              schema:
                type: string
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "429":
          $ref: "#/components/responses/Problem429"
      x-required-roles:
        - EDITOR
    post:
      tags:
        - Editorial
      operationId: createEditorIssueSection
      summary: Create an issue section
      description: Creates a section and persists its explicit position inside a draft issue.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueSectionInput"
            examples:
              valid:
                value:
                  title: 場邊現場
      responses:
        "201":
          description: Section created and ordering persisted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueSectionCollection"
              examples:
                valid:
                  value:
                    issueId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    issueVersion: 2
                    sections:
                      - sectionId: 0190f7b0-7c4b-7e3a-8f12-123456789abd
                        title: 場邊現場
                        position: 1
                        articleCount: 0
                        version: 0
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
        "429":
          $ref: "#/components/responses/Problem429"
      x-idempotent: true
      x-optimistic-lock: If-Match
      x-required-roles:
        - EDITOR
    patch:
      tags:
        - Editorial
      operationId: reorderEditorIssueSections
      summary: Reorder issue sections
      description: Atomically replaces the complete section order with contiguous explicit positions.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueSectionReorder"
            examples:
              valid:
                value:
                  sections:
                    - sectionId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                      position: 1
      responses:
        "200":
          description: Section ordering persisted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueSectionCollection"
              examples:
                valid:
                  value:
                    issueId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    issueVersion: 4
                    sections:
                      - sectionId: 0190f7b0-7c4b-7e3a-8f12-123456789abe
                        title: 人物與方法
                        position: 1
                        articleCount: 1
                        version: 1
                      - sectionId: 0190f7b0-7c4b-7e3a-8f12-123456789abd
                        title: 場邊現場
                        position: 2
                        articleCount: 2
                        version: 1
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
        "429":
          $ref: "#/components/responses/Problem429"
      x-idempotent: true
      x-optimistic-lock: If-Match
      x-required-roles:
        - EDITOR
  /api/v1/editor/issues/{issueId}/sections/{sectionId}:
    parameters:
      - $ref: "#/components/parameters/IssueId"
      - $ref: "#/components/parameters/SectionId"
    patch:
      tags:
        - Editorial
      operationId: patchEditorIssueSection
      summary: Rename an issue section
      description: Updates section metadata inside a draft issue with the issue aggregate lock.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueSectionPatch"
            examples:
              valid:
                value:
                  title: 人物與方法｜修訂
      responses:
        "200":
          description: Section metadata persisted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueSectionCollection"
              examples:
                valid:
                  value:
                    issueId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    issueVersion: 5
                    sections:
                      - sectionId: 0190f7b0-7c4b-7e3a-8f12-123456789abe
                        title: 人物與方法｜修訂
                        position: 1
                        articleCount: 1
                        version: 2
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
        "429":
          $ref: "#/components/responses/Problem429"
      x-idempotent: true
      x-optimistic-lock: If-Match
      x-required-roles:
        - EDITOR
    delete:
      tags:
        - Editorial
      operationId: deleteEditorIssueSection
      summary: Delete an empty issue section
      description: Deletes only an empty section and normalizes the remaining explicit positions.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Section deleted and ordering persisted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueSectionCollection"
              examples:
                valid:
                  value:
                    issueId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    issueVersion: 6
                    sections: []
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
        "429":
          $ref: "#/components/responses/Problem429"
      x-idempotent: true
      x-optimistic-lock: If-Match
      x-required-roles:
        - EDITOR
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC issuer/audience validated JWT for Spring Security.
    bffSession:
      type: apiKey
      in: cookie
      name: __Host-courtside_session
      description: Secure, HttpOnly, SameSite BFF session cookie.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by the previous page.
      schema:
        type: string
        minLength: 1
        maxLength: 256
    Limit:
      name: limit
      in: query
      required: false
      description: Bounded page size.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    IssueId:
      name: issueId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
    SectionId:
      name: sectionId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Optimistic-lock version or ETag. The server rejects stale values with 409 VERSION_CONFLICT.
      schema:
        type: string
        minLength: 1
        maxLength: 200
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable retry key. Replays return the original operation result.
      schema:
        type: string
        format: uuid
        minLength: 36
        maxLength: 200
  headers:
    XRequestId:
      description: Trace/request identifier. Never contains tokens or article content.
      schema:
        type: string
        minLength: 1
        maxLength: 100
...
    RetryAfter:
      description: Seconds until the caller may retry.
      schema:
        type: integer
        minimum: 1
        maximum: 3600
  schemas:
    Uuid:
      type: string
      format: uuid
      description: Stable public identifier.
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - detail
        - instance
        - requestId
        - code
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          minLength: 1
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          minLength: 1
          maxLength: 1000
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
          minLength: 1
          maxLength: 100
        code:
          type: string
          enum:
            - INVALID_REQUEST
            - AUTHENTICATION_REQUIRED
            - FORBIDDEN
            - RESOURCE_NOT_FOUND
            - VERSION_CONFLICT
            - RIGHTS_OR_CONTENT_GATE
            - RATE_LIMITED
        errors:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/FieldError"
...
    IssueDraftInput:
      type: object
      additionalProperties: false
      required:
        - title
        - slug
        - coverAssetId
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
        description:
          type: string
          maxLength: 1000
        coverAssetId:
          $ref: "#/components/schemas/Uuid"
    IssuePatch:
      type: object
      additionalProperties: false
      required:
        - issueId
        - changes
      properties:
        issueId:
          $ref: "#/components/schemas/Uuid"
        changes:
          type: object
          minProperties: 1
          additionalProperties: true
    IssueDraft:
      type: object
      additionalProperties: false
      required:
        - issueId
        - issueNumber
        - version
        - title
        - slug
        - description
        - coverAssetId
        - state
      properties:
        issueId:
          $ref: "#/components/schemas/Uuid"
        issueNumber:
          type: integer
          minimum: 1
        version:
          type: integer
          minimum: 1
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        coverAssetId:
          $ref: "#/components/schemas/Uuid"
        state:
          type: string
          enum:
            - DRAFT
            - IN_REVIEW
            - APPROVED
            - SCHEDULED
            - PUBLISHED
            - ARCHIVED
    IssueDraftPage:
      type: object
      additionalProperties: false
      required:
        - items
        - page
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/IssueDraft"
        page:
          $ref: "#/components/schemas/PageMeta"
    IssueSectionInput:
      type: object
      additionalProperties: false
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 250
        position:
          type: integer
          minimum: 1
          maximum: 50
    IssueSectionPatch:
      type: object
      additionalProperties: false
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 250
    IssueSectionReorder:
      type: object
      additionalProperties: false
      required:
        - sections
      properties:
        sections:
          type: array
          maxItems: 50
          items:
            type: object
            additionalProperties: false
            required:
              - sectionId
              - position
            properties:
              sectionId:
                $ref: "#/components/schemas/Uuid"
              position:
                type: integer
                minimum: 1
                maximum: 50
...
    IssueSectionCollection:
      type: object
      additionalProperties: false
      required:
        - issueId
        - issueVersion
        - sections
      properties:
        issueId:
          $ref: "#/components/schemas/Uuid"
        issueVersion:
          type: integer
          minimum: 1
        sections:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/IssueSection"
...
  responses:
    Problem400:
      description: Invalid request (400).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: INVALID_REQUEST example
              value:
                type: https://courtside.tw/problems/invalid_request
                title: Invalid request
                status: 400
                detail: The request is invalid.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: INVALID_REQUEST
      x-stable-error-code: INVALID_REQUEST
    Problem401:
      description: Authentication required (401).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: AUTHENTICATION_REQUIRED example
              value:
                type: https://courtside.tw/problems/authentication_required
                title: Authentication required
                status: 401
                detail: Authentication is required.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: AUTHENTICATION_REQUIRED
      x-stable-error-code: AUTHENTICATION_REQUIRED
    Problem403:
      description: Forbidden (403).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: FORBIDDEN example
              value:
                type: https://courtside.tw/problems/forbidden
                title: Forbidden
                status: 403
                detail: The actor is not allowed to perform this operation.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: FORBIDDEN
      x-stable-error-code: FORBIDDEN
    Problem404:
      description: Not found (404).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: RESOURCE_NOT_FOUND example
              value:
                type: https://courtside.tw/problems/resource_not_found
                title: Not found
                status: 404
                detail: The requested resource was not found.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: RESOURCE_NOT_FOUND
      x-stable-error-code: RESOURCE_NOT_FOUND
    Problem409:
      description: Conflict (409).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: VERSION_CONFLICT example
              value:
                type: https://courtside.tw/problems/version_conflict
                title: Conflict
                status: 409
                detail: The resource version is stale or the state conflicts.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: VERSION_CONFLICT
      x-stable-error-code: VERSION_CONFLICT
    Problem422:
      description: Unprocessable content (422).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: RIGHTS_OR_CONTENT_GATE example
              value:
                type: https://courtside.tw/problems/rights_or_content_gate
                title: Unprocessable content
                status: 422
                detail: The content or rights gate blocks this operation.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: RIGHTS_OR_CONTENT_GATE
      x-stable-error-code: RIGHTS_OR_CONTENT_GATE
    Problem429:
      description: Too many requests (429).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: RATE_LIMITED example
              value:
                type: https://courtside.tw/problems/rate_limited
                title: Too many requests
                status: 429
                detail: The request rate is limited.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: RATE_LIMITED
      x-stable-error-code: RATE_LIMITED
...